meta: Some refactoring here and there.

+++ sources/libraries/libsystem/Time.h
@@ -1,10 +1,14 @@
 #pragma once
 
+/* Copyright © 2019-2020 N. Van Bossuyt.                                      */
+/* This code is licensed under the 3-Clause BSD License.                      */
+/* See: LICENSE.md                                                            */
+
 #include <libruntime/Types.h>
 
 namespace libsystem
 {
 
-typedef uint32_t Millisecond;
+typedef uint64_t Millisecond;
 
 } // namespace libsystem
